Updates to fix a majority of warnings

+++ src/api/pdc_transform_support.c
@@ -86,7 +86,7 @@ PDCobj_transform_register(char *func, pdcid_t obj_id, int current_state, int nex
     if ((thisFtn = PDC_MALLOC(struct region_transform_ftn_info)) == NULL)
         PGOTO_ERROR(FAIL,"PDC register_obj_transforms memory allocation failed\n");
 
-    thisFtn->ftnPtr = (int (*)()) ftnPtr;
+    thisFtn->ftnPtr = (size_t (*)()) ftnPtr;
     thisFtn->object_id = obj_id;
     thisFtn->op_type = op_type;
     thisFtn->when = when;
@@ -150,7 +150,7 @@ PDCbuf_map_transform_register(char *func, void *buf, pdcid_t src_region_id, pdci
     if ((thisFtn = PDC_MALLOC(struct region_transform_ftn_info)) == NULL)
         PGOTO_ERROR(FAIL,"PDC register_obj_transforms memory allocation failed\n");
 
-    thisFtn->ftnPtr = (int (*)()) ftnPtr;
+    thisFtn->ftnPtr = (size_t (*)()) ftnPtr;
     thisFtn->object_id = dest_object_id;
     id_info = pdc_find_id(src_region_id);
     if (id_info && ((region_info = (struct PDC_region_info *)id_info->obj_ptr) != NULL))
